change: base styled components place
