+++ src/components/styled/Name.sc.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const Name = styled.h2`
+    margin: 5px 0 0;
+    font-size: 35px;
+    font-weight: 600;
+    color: #222f5c;
+`;
