+++ src/components/styled/Text.sc.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const Text = styled.p`
+    margin-top: 1em;
+    font-size: 18px;
+    font-weight: 500;
+    color: #5e6788;
+    text-align: left;
+    line-height: 1.5em;
+`;
