+++ src/components/styled/Title.sc.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const Title = styled.h1`
+    color: #7d7d7d;
+    font-size: 22px;
+    font-weight: 400;
+    text-transform: uppercase;
+    margin: 0;
+`;
