+++ src/components/styled/index.ts
@@ -1,0 +1,4 @@
+export * from "./Contacts.sc";
+export * from "./Name.sc";
+export * from "./Text.sc";
+export * from "./Title.sc";
